# packages/api-contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Daily Assistant API
  version: 0.1.0
  summary: Daily Assistant V1 API contract
  description: >-
    WP2 implements authentication, account lifecycle, capacity, invites, and admin
    operations. WP3 implements finance. WP4 implements drafts, shortcuts, and
    attachments. WP5 implements calendar events, tasks, and reminders. Trips and
    sync are implemented by WP6 and WP7. IDs are strings, timestamps are ISO 8601,
    and money is a fixed-point decimal string.
servers:
  - url: /api/v1
    description: Versioned API base path
security:
  - accessToken: []
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      summary: Get the current account summary
      tags: [Account]
      responses:
        "200":
          description: Current account summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserSummary" }
        default: { $ref: "#/components/responses/ApiError" }
  /me/close:
    post:
      operationId: closeCurrentAccount
      summary: Close the current account and release capacity
      tags: [Account]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CloseAccountRequest" }
      x-error-codes: [INVALID_CREDENTIALS]
      responses:
        "204": { description: Account closed }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        default: { $ref: "#/components/responses/ApiError" }
  /me/request-deletion:
    post:
      operationId: requestCurrentAccountDeletion
      summary: Request delayed deletion of the current account
      tags: [Account]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestDeletionRequest" }
      x-error-codes: [INVALID_CREDENTIALS]
      responses:
        "202": { description: Deletion requested and sessions revoked }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        default: { $ref: "#/components/responses/ApiError" }
  /me/sessions:
    delete:
      operationId: revokeAllSessions
      summary: Revoke every session owned by the current user
      tags: [Account]
      responses:
        "204": { description: All sessions revoked }
        default: { $ref: "#/components/responses/ApiError" }
  /me/sessions/{sessionId}:
    delete:
      operationId: revokeSession
      summary: Revoke one session owned by the current user
      tags: [Account]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204": { description: Session revoked }
        "404": { $ref: "#/components/responses/ResourceNotFound" }
  /me/change-password:
    post:
      operationId: changeCurrentPassword
      summary: Change the current password and clear the forced-change flag
      tags: [Account]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      x-error-codes: [INVALID_CURRENT_PASSWORD, PASSWORD_CHANGE_REQUIRED]
      responses:
        "204": { description: Password changed and forced-change flag cleared }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        default: { $ref: "#/components/responses/ApiError" }
...
  /admin/users:
    get:
      operationId: listAdminUsers
      summary: List account metadata without private user content
      tags: [Admin]
      responses:
        "200":
          description: Account metadata without user content
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AdminUserSummary" }
    post:
      operationId: createAdminUser
      summary: Create a user account as an administrator
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminCreateUserRequest" }
      x-error-codes:
        [CAPACITY_REACHED, DUPLICATE_RESOURCE, ADMIN_ACTION_REQUIRES_REASON]
      responses:
        "201":
          description: User created with forced password change on first login
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserSummary" }
        "400": { $ref: "#/components/responses/ApiError" }
        "409": { $ref: "#/components/responses/ApiError" }
  /admin/users/{id}/reset-password:
    post:
      operationId: resetAdminUserPassword
      summary: Reset a user password and force a change on next login
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminResetPasswordRequest" }
      x-error-codes: [ADMIN_ACTION_REQUIRES_REASON]
      responses:
        "204": { description: Password reset and forced change enabled }
        "400": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ResourceNotFound" }
  /admin/users/{id}/suspend:
    post:
      operationId: suspendAdminUser
      summary: Suspend an account without releasing capacity
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReasonRequest" }
      responses:
        "204": { description: Account suspended }
        "400": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ResourceNotFound" }
  /admin/users/{id}/close:
    post:
      operationId: closeAdminUser
      summary: Close an account and release capacity
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReasonRequest" }
      responses:
        "204": { description: Account closed }
        "400": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ResourceNotFound" }
  /admin/users/{id}/reopen:
    post:
      operationId: reopenAdminUser
      summary: Reopen an account after rechecking global capacity
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReasonRequest" }
      x-error-codes: [ADMIN_ACTION_REQUIRES_REASON]
      responses:
        "204": { description: Account reopened }
        "400": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ResourceNotFound" }
  /admin/settings:
    get:
      operationId: getAdminSystemSettings
      summary: Get capacity settings
      tags: [Admin]
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemSettings" }
    patch:
      operationId: updateAdminSystemSettings
      summary: Update capacity settings with audit
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSystemSettingsRequest" }
      x-error-codes: [SETTING_LOWER_THAN_USAGE]
      responses:
        "200":
          description: System settings updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemSettings" }
        "400": { $ref: "#/components/responses/ApiError" }
        "409": { $ref: "#/components/responses/ApiError" }
...
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token; refresh secrets are never stored in JavaScript storage.
    refreshCookie:
      type: apiKey
      in: cookie
      name: da_refresh
      description: HttpOnly, Secure in production, SameSite=Lax, rotated refresh-session cookie.
    shortcutToken:
      type: http
      scheme: bearer
      description: Revocable device credential with endpoint-specific least-privilege scopes.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identifier" }
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identifier" }
...
  responses:
    ApiError:
      description: Structured, non-sensitive API error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ResourceNotFound:
      description: Resource does not exist or is outside the current user scope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            code: RESOURCE_NOT_FOUND
            message: Resource not found
            requestId: req_example
  schemas:
    Identifier:
      type: string
      minLength: 1
      examples: ["01JEXAMPLE0000000000000000"]
    IsoDateTime:
      type: string
      format: date-time
      examples: ["2026-08-05T01:00:00.000Z"]
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          $ref: "#/components/schemas/ApiErrorCode"
        message: { type: string }
        requestId: { type: string, minLength: 1 }
        fieldErrors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    ApiErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - RATE_LIMITED
        - CAPACITY_REACHED
        - INVALID_CREDENTIALS
        - INVALID_CURRENT_PASSWORD
        - PASSWORD_CHANGE_REQUIRED
        - ACCOUNT_NOT_ACTIVE
        - REFRESH_TOKEN_REQUIRED
        - REFRESH_TOKEN_INVALID
        - INVALID_STATE
        - SETTING_LOWER_THAN_USAGE
        - ADMIN_ACTION_REQUIRES_REASON
        - IDEMPOTENCY_CONFLICT
        - VERSION_CONFLICT
        - DUPLICATE_RESOURCE
        - DRAFT_CONFIRMATION_REQUIRED
        - CREDENTIAL_INVALID
        - CREDENTIAL_REVOKED
        - ATTACHMENT_TYPE_NOT_ALLOWED
        - ATTACHMENT_TOO_LARGE
        - DRAFT_NOT_EDITABLE
        - UPLOAD_INTENT_EXPIRED
        - UPLOAD_TOKEN_INVALID
        - CONFIRMATION_TOKEN_INVALID
        - CONFIRMATION_TOKEN_EXPIRED
        - CURSOR_INVALID
        - MUTATION_BATCH_TOO_LARGE
        - MUTATION_UNSUPPORTED
        - RESOURCE_NOT_FOUND
        - INTERNAL_ERROR
...
    Password:
      type: string
      format: password
      minLength: 12
      maxLength: 128
    DisplayName:
      type: string
      minLength: 1
      maxLength: 60
    Reason:
      type: string
      minLength: 3
      maxLength: 500
    Username:
      type: string
      minLength: 3
      maxLength: 32
      pattern: "^[a-z0-9_]{3,32}$"
    UserSummary:
      type: object
      additionalProperties: false
      required:
        [
          id,
          displayName,
          username,
          role,
          status,
          createdAt,
          updatedAt,
          closedAt,
          deletionRequestedAt,
        ]
      properties:
        id: { $ref: "#/components/schemas/Identifier" }
        displayName: { $ref: "#/components/schemas/DisplayName" }
        username: { $ref: "#/components/schemas/Username" }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
        updatedAt: { $ref: "#/components/schemas/IsoDateTime" }
        closedAt:
          anyOf:
            - { $ref: "#/components/schemas/IsoDateTime" }
            - { type: "null" }
        deletionRequestedAt:
          anyOf:
            - { $ref: "#/components/schemas/IsoDateTime" }
            - { type: "null" }
...
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { $ref: "#/components/schemas/Password" }
        newPassword: { $ref: "#/components/schemas/Password" }
    CloseAccountRequest:
      type: object
      additionalProperties: false
      required: [password, reason]
      properties:
        password: { $ref: "#/components/schemas/Password" }
        reason: { $ref: "#/components/schemas/Reason" }
    RequestDeletionRequest:
      type: object
      additionalProperties: false
      required: [password, reason]
      properties:
        password: { $ref: "#/components/schemas/Password" }
        reason: { $ref: "#/components/schemas/Reason" }
    AdminReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { $ref: "#/components/schemas/Reason" }
    AdminCreateUserRequest:
      type: object
      additionalProperties: false
      required: [displayName, username, reason]
      properties:
        displayName: { $ref: "#/components/schemas/DisplayName" }
        username: { $ref: "#/components/schemas/Username" }
        initialPassword:
          anyOf:
            - { $ref: "#/components/schemas/Password" }
            - { type: "null" }
        reason: { $ref: "#/components/schemas/Reason" }
    AdminResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [newPassword, reason]
      properties:
        newPassword: { $ref: "#/components/schemas/Password" }
        reason: { $ref: "#/components/schemas/Reason" }
    SystemSettings:
      type: object
      additionalProperties: false
      required: [maxActiveUsers]
      properties:
        maxActiveUsers: { type: integer, minimum: 1, maximum: 1000 }
    UpdateSystemSettingsRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        maxActiveUsers: { type: integer, minimum: 1, maximum: 1000 }
        reason: { $ref: "#/components/schemas/Reason" }
...
    AdminUserSummary:
      type: object
      additionalProperties: false
      required:
        [
          id,
          username,
          displayName,
          mustChangePassword,
          role,
          status,
          createdAt,
          closedAt,
          deletionRequestedAt,
        ]
      properties:
        id: { $ref: "#/components/schemas/Identifier" }
        username: { $ref: "#/components/schemas/Username" }
        displayName: { $ref: "#/components/schemas/DisplayName" }
        mustChangePassword: { type: boolean }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
        closedAt:
          anyOf:
            - { $ref: "#/components/schemas/IsoDateTime" }
            - { type: "null" }
        deletionRequestedAt:
          anyOf:
            - { $ref: "#/components/schemas/IsoDateTime" }
            - { type: "null" }
...
    UserRole:
      type: string
      enum: [USER, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, CLOSED, DELETION_PENDING, DELETED]
...